requirements.txt
streamlit 
langchain-nvidia-ai-endpoints 
langchain-core
python-dotenv
tiktoken